sandbox/sandbox/swagger.py
"""Swagger API documentation."""
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from rest_framework import permissions

SchemaView = get_schema_view(
    openapi.Info(
        title="Sandbox API",
        default_version='v1',
        description="Sendbox API documentation."
    ),
    validators=['ssv', 'flex'],
    public=True,
    permission_classes=(permissions.AllowAny,)
)
