requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.5.2               # via kombu
attrs==19.3.0             # via jsonschema
billiard==3.6.1.0         # via celery
celery==4.3.0
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-appconf==1.0.3     # via django-celery-email
django-celery-email==2.0.2
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0
factory-boy==2.12.0
faker==2.0.4              # via factory-boy
idna==2.8                 # via requests
importlib-metadata==0.23  # via jsonschema, kombu
inflection==0.3.1         # via drf-yasg
invoke==1.3.0
itypes==1.1.0             # via coreapi
jinja2==2.10.3            # via coreschema
jsonschema==3.2.0         # via swagger-spec-validator
kombu==4.6.6              # via celery
markupsafe==1.1.1         # via jinja2
more-itertools==7.2.0     # via zipp
oauthlib==3.1.0           # via django-oauth-toolkit
packaging==19.2           # via drf-yasg
psycopg2-binary==2.8.4
pyparsing==2.4.5          # via packaging
pyrsistent==0.15.5        # via jsonschema
python-dateutil==2.8.1    # via faker
pytz==2019.3              # via celery, django
pyyaml==5.1.2             # via swagger-spec-validator
requests==2.22.0          # via coreapi, django-oauth-toolkit
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.5       # via drf-yasg
six==1.13.0               # via django-appconf, drf-yasg, faker, jsonschema, packaging, pyrsistent, python-dateutil, swagger-spec-validator
sqlparse==0.3.0           # via django
swagger-spec-validator==2.4.3
text-unidecode==1.3       # via faker
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.25.7           # via requests
vine==1.3.0               # via amqp, celery
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.1        # via jsonschema
